refactor(internal): move external service tokens to users.go
