+++ internal/endpoints/users.go
@@ -285,3 +285,76 @@ func setAuthCookies(w http.ResponseWriter, sessionId string, tokenInfo *users.To
 		Domain:   tokenInfo.ServiceDomain,
 	})
 }
+
+type v1ServiceUserTokens struct {
+	Code         string `json:"code,omitempty"`
+	SessionId    string `json:"session_id,omitempty"`
+	RefreshToken string `json:"refresh_token,omitempty"`
+}
+
+func (r *v1ServiceUserTokens) Valid(ctx context.Context) (problems map[string]string) {
+	problems = make(map[string]string)
+
+	usingCode := r.Code != ""
+	usingRefresh := r.SessionId != "" || r.RefreshToken != ""
+
+	switch {
+	case usingCode && usingRefresh:
+		problems["code"] = "cannot provide both 'code' and 'session_id'/'refresh_token'"
+	case !usingCode && !usingRefresh:
+		problems["code"] = "either 'code' or 'session_id' and 'refresh_token' must be provided"
+	case usingRefresh:
+		if r.SessionId == "" {
+			problems["session_id"] = "'session_id' cannot be empty when using refresh_token flow"
+		}
+		if r.RefreshToken == "" {
+			problems["refresh_token"] = "'refresh_token' cannot be empty when using refresh_token flow"
+		}
+	}
+
+	return problems
+}
+
+func V1_Service_UserTokens(logger *slog.Logger, userService *users.UserService) http.HandlerFunc {
+	type response struct {
+		AccessToken         string `json:"access_token"`
+		RefreshToken        string `json:"refresh_token"`
+		SessionId           string `json:"session_id"`
+		AccessTokenDuration int    `json:"access_token_duration_seconds"`
+	}
+
+	return func(w http.ResponseWriter, r *http.Request) {
+		data, err := decodeValid[*v1ServiceUserTokens](r)
+		if err != nil {
+			handleErr(w, r, logger, err)
+			return
+		}
+
+		var tokenInfo *users.TokenInfo
+
+		if data.Code != "" {
+			info, err := userService.CreateAccessTokenUsingCode(r.Context(), data.Code)
+			if err != nil {
+				handleErr(w, r, logger, err)
+				return
+			}
+
+			tokenInfo = info
+		} else {
+			info, err := userService.CreateAccessToken(r.Context(), data.SessionId, data.RefreshToken)
+			if err != nil {
+				handleErr(w, r, logger, err)
+				return
+			}
+
+			tokenInfo = info
+		}
+
+		safeEncode(w, r, logger, http.StatusOK, response{
+			AccessToken:         tokenInfo.AccessToken,
+			RefreshToken:        tokenInfo.RefreshToken,
+			SessionId:           tokenInfo.SessionId,
+			AccessTokenDuration: int(tokenInfo.AccessTokenDuration.Seconds()),
+		})
+	}
+}
